CLN: move test_arithmetic_explicit_conversions to common.py::NumericBase
